KNN using N-Fold Cross Validation committed
KNN is a simple clustering technique which is used for classification. I have used the famous IRIS flower Dataset to demonstrate the algorithm. This program uses a special technique called N-fold Cross Validation which is often used with KNN wherein the entire dataset is used for training and validation to instead of dividing it into train and test set as a result better accuracy is obtain and an obtimal value of K is obtained